@@ -1,38 +1,27 @@
-import App from "./components/App";
-import { AppContainer } from "react-hot-loader";
-import { initializeIcons } from "@fluentui/font-icons-mdl2";
-import { ThemeProvider } from "@fluentui/react";
 import * as React from "react";
-import * as ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
+import App from "./components/App";
+import { FluentProvider, webLightTheme } from "@fluentui/react-components";
 
 /* global document, Office, module, require */
 
-initializeIcons();
-
-let isOfficeInitialized = false;
-
 const title = "Contoso Task Pane Add-in";
 
-const render = (Component) => {
-  ReactDOM.render(
-    <AppContainer>
-      <ThemeProvider>
-        <Component title={title} isOfficeInitialized={isOfficeInitialized} />
-      </ThemeProvider>
-    </AppContainer>,
-    document.getElementById("container")
-  );
-};
+const rootElement: HTMLElement = document.getElementById("container");
+const root = createRoot(rootElement);
 
 /* Render application after Office initializes */
 Office.onReady(() => {
-  isOfficeInitialized = true;
-  render(App);
+  root.render(
+    <FluentProvider theme={webLightTheme}>
+      <App title={title} />
+    </FluentProvider>
+  );
 });
 
 if ((module as any).hot) {
   (module as any).hot.accept("./components/App", () => {
     const NextApp = require("./components/App").default;
-    render(NextApp);
+    root.render(NextApp);
   });
 }
