@@ -12,13 +12,13 @@ async function getHttpsOptions() {
 }
 
 module.exports = async (env, options) => {
-  // const dev = options.mode === "development";
+
   const config = {
     devtool: "source-map",
     entry: {
       polyfill: ["core-js/stable", "regenerator-runtime/runtime"],
-      vendor: ["react", "react-dom", "core-js", "@fluentui/react"],
-      taskpane: ["react-hot-loader/patch", path.resolve(__dirname, "./src/test.index.tsx"), path.resolve(__dirname, "./src/test-taskpane.html")],
+      vendor: ["react", "react-dom", "core-js", "@fluentui/react-components", "@fluentui/react-icons"],
+      taskpane: [path.resolve(__dirname, "./src/test.index.tsx"), path.resolve(__dirname, "./src/test-taskpane.html")],
     },
     output: {
       path: path.resolve(__dirname, "testBuild"),
@@ -46,7 +46,7 @@ module.exports = async (env, options) => {
         },
         {
           test: /\.tsx?$/,
-          use: ["react-hot-loader/webpack", "ts-loader"],
+          use: ["ts-loader"],
           exclude: /node_modules/,
         },
         {
