@@ -1,42 +1,44 @@
 import * as assert from "assert";
 import "mocha";
 import { OfficeMockObject } from "office-addin-mock";
-import * as excel from "../../src/taskpane/components/excel.App";
+import insertText from "../../src/taskpane/excel-office-document";
 
 /* global describe, global, it */
 
 const ExcelMockData = {
   context: {
     workbook: {
-      range: {
-        address: "G4",
-        format: {
-          fill: {},
+      worksheets: {
+        range: {
+          values: [[" "]],
+          format: {
+            autoFitColumns: function () { }
+          },
         },
-      },
-      getSelectedRange: function () {
-        return this.range;
-      },
-    },
+        getRange: function () {
+          return this.range;
+        },
+        getActiveWorksheet: function () {
+          return this;
+        }
+      }
+    }
   },
   run: async function (callback) {
     await callback(this.context);
   },
 };
 
-const OfficeMockData = {
-  onReady: async function () {},
-};
-
 describe("Excel", function () {
-  it("Run", async function () {
+  it("Inserts text", async function () {
     const excelMock: OfficeMockObject = new OfficeMockObject(ExcelMockData); // Mocking the host specific namespace
     global.Excel = excelMock as any;
-    global.Office = new OfficeMockObject(OfficeMockData) as any; // Mocking the common office-js namespace
 
-    const excelApp = new excel.default(this.props, this.context);
-    await excelApp.click();
+    await insertText("Hello Excel");
 
-    assert.strictEqual(excelMock.context.workbook.range.format.fill.color, "yellow");
+    excelMock.context.workbook.worksheets.range.load("values");
+    await excelMock.context.sync();
+
+   assert.strictEqual(excelMock.context.workbook.worksheets.range.values[0][0], "Hello Excel");
   });
 });
